+++ core/jobs_registry.py
@@ -25,6 +25,7 @@ from core.domain import email_jobs_one_off
 from core.domain import exp_jobs_one_off
 from core.domain import feedback_jobs_continuous
 from core.domain import feedback_jobs_one_off
+from core.domain import interaction_validation_jobs_one_off
 from core.domain import opportunity_jobs_one_off
 from core.domain import prod_validation_jobs_one_off
 from core.domain import question_jobs_one_off
@@ -49,7 +50,6 @@ ONE_OFF_JOB_MANAGERS = [
     activity_jobs_one_off.RemoveCommitUsernamesOneOffJob,
     collection_jobs_one_off.CollectionMigrationOneOffJob,
     email_jobs_one_off.EmailHashRegenerationOneOffJob,
-    exp_jobs_one_off.DragAndDropSortInputInteractionOneOffJob,
     exp_jobs_one_off.ExplorationContentValidationJobForCKEditor,
     exp_jobs_one_off.ExplorationFirstPublishedOneOffJob,
     exp_jobs_one_off.ExplorationMathTagValidationOneOffJob,
@@ -59,10 +59,6 @@ ONE_OFF_JOB_MANAGERS = [
     exp_jobs_one_off.ExplorationMigrationJobManager,
     exp_jobs_one_off.ExplorationValidityJobManager,
     exp_jobs_one_off.HintsAuditOneOffJob,
-    exp_jobs_one_off.InteractionCustomizationArgsValidationJob,
-    exp_jobs_one_off.ItemSelectionInteractionOneOffJob,
-    exp_jobs_one_off.MathExpressionValidationOneOffJob,
-    exp_jobs_one_off.MultipleChoiceInteractionOneOffJob,
     exp_jobs_one_off.ViewableExplorationsAuditJob,
     feedback_jobs_one_off.FeedbackThreadCacheOneOffJob,
     opportunity_jobs_one_off.ExplorationOpportunitySummaryModelRegenerationJob,
@@ -98,6 +94,15 @@ ONE_OFF_JOB_MANAGERS = [
 # List of all manager classes for prod validation one-off batch jobs for which
 # to show controls on the admin dashboard.
 AUDIT_JOB_MANAGERS = [
+    (
+        interaction_validation_jobs_one_off
+        .DragAndDropSortInputInteractionOneOffJob),
+    (
+        interaction_validation_jobs_one_off
+        .InteractionRTECustomizationArgsValidationJob),
+    interaction_validation_jobs_one_off.ItemSelectionInteractionOneOffJob,
+    interaction_validation_jobs_one_off.MathExpressionValidationOneOffJob,
+    interaction_validation_jobs_one_off.MultipleChoiceInteractionOneOffJob,
     prod_validation_jobs_one_off.ActivityReferencesModelAuditOneOffJob,
     prod_validation_jobs_one_off.BulkEmailModelAuditOneOffJob,
     prod_validation_jobs_one_off.ClassifierTrainingJobModelAuditOneOffJob,
